backend/requirements.txt
# Dépendances principales
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.9

# Traitement de documents PDF
PyMuPDF>=1.23.8  # Remplace pymupdf
pdf2image>=1.16.3
PyPDF2>=3.0.1
poppler-utils>=0.1.0  # à installer via pip ou système

# Vision par ordinateur et traitement d'images
opencv-python>=4.8.1.78
Pillow>=11.0.0
numpy>=1.26.0

# Traitement de données et science des données
pandas>=2.2.0
scikit-learn>=1.3.0

# Web scraping et HTTP
requests>=2.31.0
beautifulsoup4>=4.12.2

# Intelligence artificielle et ML
huggingface-hub>=0.19.0
transformers>=4.35.0
sentencepiece>=0.1.99
torch>=2.1.0
gpt4all>=1.0.0

# OCR
pytesseract>=0.3.10

# Web et API
gradio_client>=0.7.0

# Utilitaires
tqdm>=4.66.1
streamlit>=1.28.0
psutil>=5.9.0

# Tests
pytest>=7.4.0
pytest-mock>=3.11.1
pytest-cov>=4.1.0

# Qualité de code
black>=23.10.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.7.0
ruff>=0.1.6

# Typage
typing-extensions>=4.8.0

